plugin/s_base/requirements.txt
ShopifyAPI==8.4.*
xmltodict==0.12.*
google-api-python-client==2.2.*
customerio==1.1
pika==1.2.*